port to DPDK 18.11

+++ rte/src/bond.rs
@@ -3,9 +3,9 @@ use std::mem;
 use ffi;
 
 use errors::Result;
-use memory::SocketId;
 use ethdev;
 use ether;
+use memory::SocketId;
 
 /// Supported modes of operation of link bonding library
 #[repr(u8)]
@@ -197,10 +197,11 @@ impl BondedDevice for ethdev::PortId {
     }
 
     fn slaves(&self) -> Result<Vec<ethdev::PortId>> {
-        let mut slaves = [0u8; ffi::RTE_MAX_ETHPORTS as usize];
+        let mut slaves = [0u16; ffi::RTE_MAX_ETHPORTS as usize];
 
-        let num =
-            unsafe { ffi::rte_eth_bond_slaves_get(*self, slaves.as_mut_ptr(), slaves.len() as u8) };
+        let num = unsafe {
+            ffi::rte_eth_bond_slaves_get(*self, slaves.as_mut_ptr(), slaves.len() as u16)
+        };
 
         rte_check!(num; ok => {
             Vec::from(&slaves[..num as usize])
@@ -208,10 +209,11 @@ impl BondedDevice for ethdev::PortId {
     }
 
     fn active_slaves(&self) -> Result<Vec<ethdev::PortId>> {
-        let mut slaves = [0u8; ffi::RTE_MAX_ETHPORTS as usize];
+        let mut slaves = [0u16; ffi::RTE_MAX_ETHPORTS as usize];
 
-        let num =
-            unsafe { ffi::rte_eth_bond_slaves_get(*self, slaves.as_mut_ptr(), slaves.len() as u8) };
+        let num = unsafe {
+            ffi::rte_eth_bond_slaves_get(*self, slaves.as_mut_ptr(), slaves.len() as u16)
+        };
 
         rte_check!(num; ok => {
             Vec::from(&slaves[..num as usize])
